修复SwiGluQuant大shape精度问题



# message auto-generated for no-merge-commit merge:
!143 merge master into master

修复SwiGluQuant大shape精度问题

Created-by: songjianquan
Commit-by: songjianquan
Merged-by: turing_project1
Description: ## 描述
修复SwiGluQuant大shape精度问题

## 关联的Issue
https://gitcode.com/cann/ops-nn/issues/67

## 测试
![image.png](https://raw.gitcode.com/user-images/assets/7665709/49b01cfd-448f-49af-bb00-c059fcf36c33/image.png 'image.png')

## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [ ] 表示选中 -->
- [x] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [ ] 其他，请描述：




See merge request: cann/ops-nn!143

+++ quant/swi_glu_quant/op_kernel/swi_glu_quant.h
@@ -164,7 +164,7 @@ private:
         }
     }
 
-    __aicore__ inline void ProcessCoreMultiUbMultiAlign(uint32_t ridx, uint32_t &smoothIndex, uint16_t offsetRow)
+    __aicore__ inline void ProcessCoreMultiUbMultiAlign(uint32_t ridx, uint32_t &smoothIndex, uint32_t offsetRow)
     {
         DataCopyParams splitCopyinParams;
         DataCopyParams splitCopyoutParams;
